Re-organize registries into sub-package-level config.py files
