+++ glue/qt/glue_application.py
@@ -588,7 +588,7 @@ class GlueApplication(Application, QMainWindow):
         self._actions['data_new'] = a
 
         # We now populate the "Import data" menu
-        from glue.config import importer
+        from ..core.config import importer
 
         acts = []
 
@@ -607,7 +607,7 @@ class GlueApplication(Application, QMainWindow):
 
         self._actions['data_importers'] = acts
 
-        from glue.config import exporters
+        from ..core.config import exporters
         if len(exporters) > 0:
             acts = []
             for e in exporters:
@@ -646,7 +646,7 @@ class GlueApplication(Application, QMainWindow):
         self._actions['redo'] = a
 
         # Create actions for menubar plugins
-        from glue.config import menubar_plugin
+        from ..core.config import menubar_plugin
         acts = []
         for label, function in menubar_plugin:
             a = act(label, self, tip=label)
@@ -665,7 +665,7 @@ class GlueApplication(Application, QMainWindow):
         """ Create a new visualization window in the current tab
         """
 
-        from ..config import qt_client
+        from .config import qt_client
         from .widgets import ScatterWidget, ImageWidget
 
         if data and data.ndim == 1 and ScatterWidget in qt_client.members:
